fix(grammar-coverage): commit only the compact baseline, and satisfy lint and the docs build
The committed artifact was every row (691K JSON plus a 149K table). It is now
counts, corpora, no-evidence productions and unseen forms only (11K), written
by -baseline; the full JSON, tables and text summary stay in build/. Also
registers the page in the mkdocs nav so the strict build passes, and clears the
gosec findings (report permissions, unneeded int conversion, #nosec on the
paths the command line names).

+++ cmd/grammar-coverage/classify.go
@@ -521,7 +521,7 @@ func newLitTable(literals []string) (*litTable, error) {
 func (t *litTable) mask(literal string) mask {
 	var m mask
 	if bit, ok := t.bit[literal]; ok {
-		m[bit/64] |= 1 << uint(bit%64)
+		m[bit/64] |= 1 << (bit % 64)
 	}
 	return m
 }
